Add checks to prevent some player stats from overflowing thus making such exploits as 65k SP impossible. This change affects such stats as spell power, feral AP, mana regen and health regen.
--HG--
branch : trunk

+++ src/server/game/Entities/Player/Player.h
@@ -2501,10 +2501,10 @@ class Player : public Unit, public GridObject<Player>
 
         float m_auraBaseMod[BASEMOD_END][MOD_END];
         int16 m_baseRatingValue[MAX_COMBAT_RATING];
-        uint16 m_baseSpellPower;
-        uint16 m_baseFeralAP;
-        uint16 m_baseManaRegen;
-        uint16 m_baseHealthRegen;
+        uint32 m_baseSpellPower;
+        uint32 m_baseFeralAP;
+        uint32 m_baseManaRegen;
+        uint32 m_baseHealthRegen;
 
         SpellModList m_spellMods[MAX_SPELLMOD];
         //uint32 m_pad;
